优化bug

+++ app/src/main/java/com/article/oa_article/module/task_allot/Task_allotFragment.java
@@ -293,6 +293,9 @@ public class Task_allotFragment extends MVPBaseFragment<Task_allotContract.View,
     private void setSwipeMenu() {
         // 创建菜单：
         SwipeMenuCreator mSwipeMenuCreator = (leftMenu, rightMenu, viewType) -> {
+            if(viewType == 3){  //已取消
+                return;
+            }
             // 2 删除
             SwipeMenuItem deleteItem = new SwipeMenuItem(getActivity());
             if (viewType == 0) {
